Correctly evaluate which sysctl files are readable when running GetPattern()
When running GetPattern(), the code considered a file readable if
the read bit of other users was set. This yielded inaccurate results,
for example in the case the owner of the sysctl file has read permissions
but other users do not.

This commit fixes the issue by trying to open the file and skipping
if the open is not successful, which is a more reliable way to
validdate wheter the file is readable.